New methods in LogHelper
Move to LogHelper.LogExceptionMessage methods

The existing LogHelper methods instantiate the exception in them, which is convenient,
but it makes it impossible to set properties on typed exceptions to give more info
to the developer. As part of moving toward more informative exceptions with properties,
methods will now create the exception locally and then call LogHelper methods to log
the exception. Existing LogHelper methods were public so they can't be removed,
but internally we'll stop using them and instead follow a pattern of instantiating the
exception locally to the calling code.